requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
transformers==4.37.2
torch==2.2.0
python-dotenv==1.0.1